pyproject.toml
[tool.poetry]
name = "Cache"
version = "0.1.0"
description = ""
authors = ["tmpuser1233 <ttmp1806@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.4"
pandas = "^2.2.1"
fire = "^0.5.0"
joblib = "^1.3.2"
conf = "^0.4.1"
matplotlib = "^3.8.3"
hydra-core = "^1.3.2"
scipy = "^1.12.0"
seaborn = "^0.13.2"
ipywidgets = "^8.1.2"
numba = "^0.59.1"
tqdm = "^4.66.2"
dvc-gdrive = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
